hype_dca/requirements.txt
hyperliquid-python-sdk>=0.9.0
web3>=6.0.0
eth-account>=0.10.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
requests>=2.31.0
